modulized the table building files, added set timeouts
